[update]ATR-ML戦略にリファクタリング済み設計を統合 #3
設計文書（issue_3_1_architecture_refactoring_design.md）に基づき、
リファクタリング済みの2層戦略アーキテクチャを元のatr_ml_strategy.pyに統合。

変更内容:
- 設定駆動型アーキテクチャの適用（環境変数・設定ファイルで戦略切替可能）
- 統合ファクトリーパターンによる1次・2次モデルの柔軟な構成
- FreqAI統合の改善（条件付き有効化、ML予測による指値価格調整）
- コードの簡潔化（統合戦略による処理の委譲）
- 一時ファイル（atr_ml_strategy_refactored.py）の削除
